UC8176 and UC8151 ePaper drivers
Support for Waveshare Pico-ePaper-4.2 and Badger 2040 (UC8151 and UC8176 drivers)
Still improving display driver architecture ..

+++ Source/Samples/MCU/8x8Matrix.hs
@@ -12,11 +12,11 @@ program LED8x8Matrix
     TestDrawRect()
     {
         Screen.Clear();
-        int ph = int(Display.PixelHeight);
-        int pw = int(Display.PixelWidth);
+        int ph = Display.PixelHeight;
+        int pw = Display.PixelWidth;
         for(int i=0; i< ph / 2; i += 1)
         {
-            Rectangle(i, i, uint(pw-2*i), uint(ph-2*i), Color.White);
+            Rectangle(i, i, pw-2*i, ph-2*i, Color.White);
             Delay(250);
         }
     }
@@ -24,12 +24,12 @@ program LED8x8Matrix
     TestFillRect()
     {
         Screen.Clear();
-        int ph = int(Display.PixelHeight);
-        int pw = int(Display.PixelWidth);
+        int ph = Display.PixelHeight;
+        int pw = Display.PixelWidth;
         for(int i=0; i<ph / 2; i += 1)
         {
             // Color.Inverse is used so rectangles alternate white/black
-            FilledRectangle(i, i, uint(pw-i*2), uint(ph-i*2), Color.Invert);
+            FilledRectangle(i, i, pw-i*2, ph-i*2, Color.Invert);
             Delay(250);
         }
     }
@@ -46,8 +46,8 @@ program LED8x8Matrix
     TestDrawLines() 
     {
         int i;
-        int ph = int(Display.PixelHeight);
-        int pw = int(Display.PixelWidth);
+        int ph = Display.PixelHeight;
+        int pw = Display.PixelWidth;
         Screen.Clear();
         for(i=0; i<pw; i += 1) 
         {
